Try to break up all tests into the chunks
