+++ .github/workflows/pd-tests.yaml
@@ -46,8 +46,7 @@ jobs:
           WORKER_ID: ${{ matrix.worker_id }}
           WORKER_COUNT: 10
         run: |
-          [ $WORKER_ID -ne $WORKER_COUNT ] && make ci-test-job JOB_COUNT=$(($WORKER_COUNT-1)) JOB_INDEX=$WORKER_ID
-          [ $WORKER_ID -eq $WORKER_COUNT ] && make ci-test-job-submod
+          make ci-test-job JOB_COUNT=$(($WORKER_COUNT)) JOB_INDEX=$WORKER_ID
           mv covprofile covprofile_$WORKER_ID
           sed -i "/failpoint_binding/d" covprofile_$WORKER_ID
       - name: Upload coverage result ${{ matrix.worker_id }}
