+++ tests/integrations/client/cert/ca.pem
