+++ tests/integrations/client/global_config_test.go
