+++ tests/integrations/tso/consistency_test.go
@@ -30,7 +30,7 @@ import (
 	pd "github.com/tikv/pd/pkg/utils/testutil"
 	"github.com/tikv/pd/pkg/utils/tsoutil"
 	"github.com/tikv/pd/tests"
-	"github.com/tikv/pd/tests/mcs"
+	"github.com/tikv/pd/tests/integrations/mcs"
 	"google.golang.org/grpc"
 )
