+++ tests/integrations/tso/server_test.go
@@ -28,7 +28,7 @@ import (
 	"github.com/tikv/pd/pkg/utils/tempurl"
 	pd "github.com/tikv/pd/pkg/utils/testutil"
 	"github.com/tikv/pd/tests"
-	"github.com/tikv/pd/tests/mcs"
+	"github.com/tikv/pd/tests/integrations/mcs"
 	"google.golang.org/grpc"
 )
